Add an event marker to a chart

Add a point with a label based on a datetime.

Usage

add_event_marker(
  ax,
  when,
  y,
  size = 5,
  color = "#000",
  fill = "#FFF",
  width = 2,
  shape = "circle",
  radius = 2,
  label = NULL,
  ...
)

Arguments

ax

An apexchart() htmlwidget object.

when

Vector of position to place the event.

y

Coordinate(s) on the y-axis.

size

Size of the marker.

color

Stroke Color of the marker point.

fill

Fill Color of the marker point.

width

Stroke Size of the marker point.

shape

Shape of the marker: "circle" or "square".

radius

Radius of the marker (applies to square shape).

label

Add a label to the shade, use a character or see label for more controls.

...

Additional arguments, see https://apexcharts.com/docs/options/annotations/ for possible options.

Value

An apexchart()

htmlwidget object.

See also

Examples

library(apexcharter)
data("consumption")

# add a marker
apex(consumption, aes(date, value, group = type), "spline") %>% 
  add_event_marker(when = "2020-01-22", y = 1805)

# with a label
apex(consumption, aes(date, value, group = type), "spline") %>% 
  add_event_marker(when = "2020-01-22", y = 1805, label = "Consumption peak")

# add several markers
apex(consumption, aes(date, value, group = type), "spline") %>% 
  add_event_marker(
    when = c("2020-01-02", "2020-01-06", "2020-01-13", 
             "2020-01-22", "2020-01-28", "2020-02-06",
             "2020-02-13", "2020-02-19", "2020-02-27"),
    y = c(1545, 1659, 1614,
          1805, 1637, 1636, 
          1597, 1547, 1631),
    size = 10,
    color = "firebrick"
  )