Chart parameters

Chart parameters

ax_chart(
  ax,
  type = NULL,
  stacked = NULL,
  stackType = NULL,
  defaultLocale = NULL,
  locales = NULL,
  animations = NULL,
  background = NULL,
  foreColor = NULL,
  dropShadow = NULL,
  events = NULL,
  offsetX = NULL,
  offsetY = NULL,
  selection = NULL,
  sparkline = NULL,
  toolbar = NULL,
  zoom = NULL,
  width = NULL,
  height = NULL,
  ...
)

Arguments

ax
A apexcharts htmlwidget object.

type
Specify the chart type. Available Options: "bar", "column", "line", "pie", "donut", "radialBar", "scatter", "bubble", "heatmap".

stacked
Logical. Enables stacked option for axis charts.

stackType
When stacked, should the stacking be percentage based or normal stacking. Available options: "normal" or "100%".

defaultLocale
Locale to use : "ca", "de", "el", "en", "es", "fi", "fr", "hi", "hr", "hy", "id", "it", "ko", "nl", "pt-br", "ru", "se", "tr", "ua".

locales
Array of custom locales parameters.

animations
A list of parameters.

background
Background color for the chart area. If you want to set background with css, use .apexcharts-canvas to set it.

foreColor
Sets the text color for the chart. Defaults to #373d3f.

dropShadow
A list of parameters. See https://apexcharts.com/docs/options/chart/dropshadow/.

events
See events_opts.

offsetX
Sets the left offset for chart.

offsetY
Sets the top offset for chart.

selection
A list of parameters.

sparkline
List. Sparkline hides all the elements of the charts other than the primary paths. Helps to visualize data in small areas. .

toolbar
A list of parameters. See https://apexcharts.com/docs/options/chart/toolbar/.

zoom
A list of parameters. See https://apexcharts.com/docs/options/chart/zoom/.

width
Width of the chart.

height
Height of the chart.

...
Additional parameters.

Value

A apexcharts htmlwidget object.

Examples

library(dplyr)
data("diamonds", package = "ggplot2")

##  Stack bar type
# default is dodge
apex(
  data = count(diamonds, cut, color),
  mapping = aes(x = cut, y = n, fill = color)
)

# stack
apex(
  data = count(diamonds, cut, color),
  mapping = aes(x = cut, y = n, fill = color)
) %>%
  ax_chart(stacked = TRUE)

# stack filled
apex(
  data = count(diamonds, cut, color),
  mapping = aes(x = cut, y = n, fill = color)
) %>%
  ax_chart(stacked = TRUE, stackType = "100%")

# Toolbar --------------------------------------

# Hide the toolbar
apex(
  data = count(diamonds, cut, color),
  mapping = aes(x = cut, y = n, fill = color)
) %>%
  ax_chart(toolbar = list(show = FALSE))

# Hide download buttons
data("economics", package = "ggplot2")
apex(
  data = economics,
  mapping = aes(x = date, y = pce),
  type = "line"
) %>%
  ax_chart(
    toolbar = list(tools= list(download = FALSE))
  )

# Zoom -----------------------------------------

# Disable
apex(
  data = economics,
  mapping = aes(x = date, y = pce),
  type = "line"
) %>%
  ax_chart(
    zoom = list(enabled = FALSE)
  )

# Auto-scale Y axis
apex(
  data = economics,
  mapping = aes(x = date, y = pce),
  type = "line"
) %>%
  ax_chart(
    zoom = list(autoScaleYaxis = TRUE)
  )

# Localization ---------------------------------

# Use included localization config
dat <- data.frame(
  x = Sys.Date() + 1:20,
  y = sample.int(20, 20)
)

# French
apex(dat, aes(x, y), "line") %>%
  ax_chart(defaultLocale = "fr")

# Italian
apex(dat, aes(x, y), "line") %>%
  ax_chart(defaultLocale = "it")

# Custom config
apex(dat, aes(x, y), "line") %>%
  ax_chart(locales = list(
    list(
      name = "en", # override 'en' locale
      options = list(
        toolbar = list(
          exportToSVG = "GET SVG",
          exportToPNG = "GET PNG"
        )
      )
    )
  ))