Bubble options

Use these options in ax_plotOptions().

Usage

bubble_opts(minBubbleRadius, maxBubbleRadius, ...)

Arguments

minBubbleRadius

Minimum radius size of a bubble. If a bubble value is too small to be displayed, this size will be used.

maxBubbleRadius

Maximum radius size of a bubble. If a bubble value is too large to cover the chart, this size will be used.

...

Additional parameters.

Value

A list of options that can be used in ax_plotOptions().

Note

See https://apexcharts.com/docs/options/plotoptions/bubble/.

Examples

apex(
  data = mtcars, 
  type = "scatter", 
  mapping = aes(x = wt, y = mpg, z = qsec)
) %>% 
  ax_plotOptions(
    bubble = bubble_opts(
      minBubbleRadius = 1,
      maxBubbleRadius = 20
    )
  )