Fixed tooltip

Fixed tooltip

Usage

set_tooltip_fixed(
  ax,
  position = c("topLeft", "topRight", "bottomLeft", "bottomRight"),
  offsetX = NULL,
  offsetY = NULL
)

Arguments

ax

An apexchart() htmlwidget object.

position

Predefined position: "topLeft", "topRight", "bottomLeft" or "bottomRight".

offsetX

Sets the left offset for the tooltip container in fixed position.

offsetY

Sets the top offset for the tooltip container in fixed position.

Value

An apexchart() htmlwidget object.

Examples

library(apexcharter)
data("economics", package = "ggplot2")

apex(
  data = tail(economics, 350), 
  mapping = aes(x = date, y = uempmed),
  type = "line"
) %>% 
  set_tooltip_fixed()