Add horizontal or vertical line

Add horizontal or vertical line

add_hline(ax, value, color = "#000", dash = 0, label = NULL, ...)

add_vline(ax, value, color = "#000", dash = 0, label = NULL, ...)

Arguments

ax
An apexcharts htmlwidget object.

value
Vector of position for the line(s).

color
Color(s) of the line(s).

dash
Creates dashes in borders of SVG path. A higher number creates more space between dashes in the border. Use 0 for plain line.

label
Add a label to the shade, use a character or see label for more controls.

...
Additional arguments, see https://apexcharts.com/docs/options/annotations/ for possible options.

Value

An apexcharts htmlwidget object.

Examples

library(apexcharter)

# On a column chart
apex(
  data = table(unhcr_popstats_2017$continent_residence),
  aes(Var1, Freq),
  "column"
) %>% 
  add_hline(value = 2100)

# On a scatter chart
apex(
  data = iris,
  aes(Sepal.Length, Sepal.Width),
  "scatter"
) %>% 
  add_hline(value = mean(iris$Sepal.Width)) %>% 
  add_vline(value = mean(iris$Sepal.Length))

# With labels
apex(
  data = iris,
  aes(Sepal.Length, Sepal.Width),
  "scatter"
) %>% 
  add_hline(
    value = mean(iris$Sepal.Width),
    label = "Mean of Sepal.Width"
  ) %>% 
  add_vline(
    value = mean(iris$Sepal.Length),
    label = "Mean of Sepal.Length"
  )