Chart's title

Chart's title

Usage

ax_title(
  ax,
  text = NULL,
  align = NULL,
  margin = NULL,
  offsetX = NULL,
  offsetY = NULL,
  floating = NULL,
  style = NULL,
  ...
)

Arguments

ax

An apexchart() htmlwidget object.

text

Text to display as a title of chart.

align

Alignment of subtitle relative to chart area. Possible Options: "left", "center" and "right".

margin

Numeric. Vertical spacing around the title text.

offsetX

Numeric. Sets the left offset for subtitle text.

offsetY

Numeric. Sets the top offset for subtitle text

floating

Logical. The floating option will take out the subtitle text from the chart area and make it float on top of the chart.

style

List with two items: fontSize (Font Size of the title text) and color (Fore color of the title text).

...

Additional parameters.

Value

An apexchart()

htmlwidget object.

Note

See https://apexcharts.com/docs/options/title/

Examples

data("economics", package = "ggplot2")
apex(
  data = economics,
  mapping = aes(x = date, y = uempmed),
  type = "line"
) %>% 
  ax_title(
    text = "Median duration of unemployment, in weeks"
  )