Convert a data.frame to a list

Convert data to a format suitable for ApexCharts.js

parse_df(data, add_names = FALSE)

Arguments

data
A data.frame or an object coercible to data.frame.

add_names
Use names of columns in output. Can be logical to reuse data names or a character vector of new names.

Value

A list that can be used to specify data in ax_series for example.

Examples

# All iris dataset
parse_df(iris)

#> [[1]]
#> [[1]][[1]]
#> [1] 5.1
#> 
#> [[1]][[2]]
#> [1] 3.5
#> 
#> [[1]][[3]]
#> [1] 1.4
#> 
#> [[1]][[4]]
#> [1] 0.2
#> 
#> [[1]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[2]]
#> [[2]][[1]]
#> [1] 4.9
#> 
#> [[2]][[2]]
#> [1] 3
#> 
#> [[2]][[3]]
#> [1] 1.4
#> 
#> [[2]][[4]]
#> [1] 0.2
#> 
#> [[2]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[3]]
#> [[3]][[1]]
#> [1] 4.7
#> 
#> [[3]][[2]]
#> [1] 3.2
#> 
#> [[3]][[3]]
#> [1] 1.3
#> 
#> [[3]][[4]]
#> [1] 0.2
#> 
#> [[3]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[4]]
#> [[4]][[1]]
#> [1] 4.6
#> 
#> [[4]][[2]]
#> [1] 3.1
#> 
#> [[4]][[3]]
#> [1] 1.5
#> 
#> [[4]][[4]]
#> [1] 0.2
#> 
#> [[4]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[5]]
#> [[5]][[1]]
#> [1] 5
#> 
#> [[5]][[2]]
#> [1] 3.6
#> 
#> [[5]][[3]]
#> [1] 1.4
#> 
#> [[5]][[4]]
#> [1] 0.2
#> 
#> [[5]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[6]]
#> [[6]][[1]]
#> [1] 5.4
#> 
#> [[6]][[2]]
#> [1] 3.9
#> 
#> [[6]][[3]]
#> [1] 1.7
#> 
#> [[6]][[4]]
#> [1] 0.4
#> 
#> [[6]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[7]]
#> [[7]][[1]]
#> [1] 4.6
#> 
#> [[7]][[2]]
#> [1] 3.4
#> 
#> [[7]][[3]]
#> [1] 1.4
#> 
#> [[7]][[4]]
#> [1] 0.3
#> 
#> [[7]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[8]]
#> [[8]][[1]]
#> [1] 5
#> 
#> [[8]][[2]]
#> [1] 3.4
#> 
#> [[8]][[3]]
#> [1] 1.5
#> 
#> [[8]][[4]]
#> [1] 0.2
#> 
#> [[8]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[9]]
#> [[9]][[1]]
#> [1] 4.4
#> 
#> [[9]][[2]]
#> [1] 2.9
#> 
#> [[9]][[3]]
#> [1] 1.4
#> 
#> [[9]][[4]]
#> [1] 0.2
#> 
#> [[9]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[10]]
#> [[10]][[1]]
#> [1] 4.9
#> 
#> [[10]][[2]]
#> [1] 3.1
#> 
#> [[10]][[3]]
#> [1] 1.5
#> 
#> [[10]][[4]]
#> [1] 0.1
#> 
#> [[10]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[11]]
#> [[11]][[1]]
#> [1] 5.4
#> 
#> [[11]][[2]]
#> [1] 3.7
#> 
#> [[11]][[3]]
#> [1] 1.5
#> 
#> [[11]][[4]]
#> [1] 0.2
#> 
#> [[11]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[12]]
#> [[12]][[1]]
#> [1] 4.8
#> 
#> [[12]][[2]]
#> [1] 3.4
#> 
#> [[12]][[3]]
#> [1] 1.6
#> 
#> [[12]][[4]]
#> [1] 0.2
#> 
#> [[12]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[13]]
#> [[13]][[1]]
#> [1] 4.8
#> 
#> [[13]][[2]]
#> [1] 3
#> 
#> [[13]][[3]]
#> [1] 1.4
#> 
#> [[13]][[4]]
#> [1] 0.1
#> 
#> [[13]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[14]]
#> [[14]][[1]]
#> [1] 4.3
#> 
#> [[14]][[2]]
#> [1] 3
#> 
#> [[14]][[3]]
#> [1] 1.1
#> 
#> [[14]][[4]]
#> [1] 0.1
#> 
#> [[14]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[15]]
#> [[15]][[1]]
#> [1] 5.8
#> 
#> [[15]][[2]]
#> [1] 4
#> 
#> [[15]][[3]]
#> [1] 1.2
#> 
#> [[15]][[4]]
#> [1] 0.2
#> 
#> [[15]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[16]]
#> [[16]][[1]]
#> [1] 5.7
#> 
#> [[16]][[2]]
#> [1] 4.4
#> 
#> [[16]][[3]]
#> [1] 1.5
#> 
#> [[16]][[4]]
#> [1] 0.4
#> 
#> [[16]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[17]]
#> [[17]][[1]]
#> [1] 5.4
#> 
#> [[17]][[2]]
#> [1] 3.9
#> 
#> [[17]][[3]]
#> [1] 1.3
#> 
#> [[17]][[4]]
#> [1] 0.4
#> 
#> [[17]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[18]]
#> [[18]][[1]]
#> [1] 5.1
#> 
#> [[18]][[2]]
#> [1] 3.5
#> 
#> [[18]][[3]]
#> [1] 1.4
#> 
#> [[18]][[4]]
#> [1] 0.3
#> 
#> [[18]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[19]]
#> [[19]][[1]]
#> [1] 5.7
#> 
#> [[19]][[2]]
#> [1] 3.8
#> 
#> [[19]][[3]]
#> [1] 1.7
#> 
#> [[19]][[4]]
#> [1] 0.3
#> 
#> [[19]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[20]]
#> [[20]][[1]]
#> [1] 5.1
#> 
#> [[20]][[2]]
#> [1] 3.8
#> 
#> [[20]][[3]]
#> [1] 1.5
#> 
#> [[20]][[4]]
#> [1] 0.3
#> 
#> [[20]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[21]]
#> [[21]][[1]]
#> [1] 5.4
#> 
#> [[21]][[2]]
#> [1] 3.4
#> 
#> [[21]][[3]]
#> [1] 1.7
#> 
#> [[21]][[4]]
#> [1] 0.2
#> 
#> [[21]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[22]]
#> [[22]][[1]]
#> [1] 5.1
#> 
#> [[22]][[2]]
#> [1] 3.7
#> 
#> [[22]][[3]]
#> [1] 1.5
#> 
#> [[22]][[4]]
#> [1] 0.4
#> 
#> [[22]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[23]]
#> [[23]][[1]]
#> [1] 4.6
#> 
#> [[23]][[2]]
#> [1] 3.6
#> 
#> [[23]][[3]]
#> [1] 1
#> 
#> [[23]][[4]]
#> [1] 0.2
#> 
#> [[23]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[24]]
#> [[24]][[1]]
#> [1] 5.1
#> 
#> [[24]][[2]]
#> [1] 3.3
#> 
#> [[24]][[3]]
#> [1] 1.7
#> 
#> [[24]][[4]]
#> [1] 0.5
#> 
#> [[24]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[25]]
#> [[25]][[1]]
#> [1] 4.8
#> 
#> [[25]][[2]]
#> [1] 3.4
#> 
#> [[25]][[3]]
#> [1] 1.9
#> 
#> [[25]][[4]]
#> [1] 0.2
#> 
#> [[25]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[26]]
#> [[26]][[1]]
#> [1] 5
#> 
#> [[26]][[2]]
#> [1] 3
#> 
#> [[26]][[3]]
#> [1] 1.6
#> 
#> [[26]][[4]]
#> [1] 0.2
#> 
#> [[26]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[27]]
#> [[27]][[1]]
#> [1] 5
#> 
#> [[27]][[2]]
#> [1] 3.4
#> 
#> [[27]][[3]]
#> [1] 1.6
#> 
#> [[27]][[4]]
#> [1] 0.4
#> 
#> [[27]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[28]]
#> [[28]][[1]]
#> [1] 5.2
#> 
#> [[28]][[2]]
#> [1] 3.5
#> 
#> [[28]][[3]]
#> [1] 1.5
#> 
#> [[28]][[4]]
#> [1] 0.2
#> 
#> [[28]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[29]]
#> [[29]][[1]]
#> [1] 5.2
#> 
#> [[29]][[2]]
#> [1] 3.4
#> 
#> [[29]][[3]]
#> [1] 1.4
#> 
#> [[29]][[4]]
#> [1] 0.2
#> 
#> [[29]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[30]]
#> [[30]][[1]]
#> [1] 4.7
#> 
#> [[30]][[2]]
#> [1] 3.2
#> 
#> [[30]][[3]]
#> [1] 1.6
#> 
#> [[30]][[4]]
#> [1] 0.2
#> 
#> [[30]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[31]]
#> [[31]][[1]]
#> [1] 4.8
#> 
#> [[31]][[2]]
#> [1] 3.1
#> 
#> [[31]][[3]]
#> [1] 1.6
#> 
#> [[31]][[4]]
#> [1] 0.2
#> 
#> [[31]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[32]]
#> [[32]][[1]]
#> [1] 5.4
#> 
#> [[32]][[2]]
#> [1] 3.4
#> 
#> [[32]][[3]]
#> [1] 1.5
#> 
#> [[32]][[4]]
#> [1] 0.4
#> 
#> [[32]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[33]]
#> [[33]][[1]]
#> [1] 5.2
#> 
#> [[33]][[2]]
#> [1] 4.1
#> 
#> [[33]][[3]]
#> [1] 1.5
#> 
#> [[33]][[4]]
#> [1] 0.1
#> 
#> [[33]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[34]]
#> [[34]][[1]]
#> [1] 5.5
#> 
#> [[34]][[2]]
#> [1] 4.2
#> 
#> [[34]][[3]]
#> [1] 1.4
#> 
#> [[34]][[4]]
#> [1] 0.2
#> 
#> [[34]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[35]]
#> [[35]][[1]]
#> [1] 4.9
#> 
#> [[35]][[2]]
#> [1] 3.1
#> 
#> [[35]][[3]]
#> [1] 1.5
#> 
#> [[35]][[4]]
#> [1] 0.2
#> 
#> [[35]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[36]]
#> [[36]][[1]]
#> [1] 5
#> 
#> [[36]][[2]]
#> [1] 3.2
#> 
#> [[36]][[3]]
#> [1] 1.2
#> 
#> [[36]][[4]]
#> [1] 0.2
#> 
#> [[36]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[37]]
#> [[37]][[1]]
#> [1] 5.5
#> 
#> [[37]][[2]]
#> [1] 3.5
#> 
#> [[37]][[3]]
#> [1] 1.3
#> 
#> [[37]][[4]]
#> [1] 0.2
#> 
#> [[37]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[38]]
#> [[38]][[1]]
#> [1] 4.9
#> 
#> [[38]][[2]]
#> [1] 3.6
#> 
#> [[38]][[3]]
#> [1] 1.4
#> 
#> [[38]][[4]]
#> [1] 0.1
#> 
#> [[38]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[39]]
#> [[39]][[1]]
#> [1] 4.4
#> 
#> [[39]][[2]]
#> [1] 3
#> 
#> [[39]][[3]]
#> [1] 1.3
#> 
#> [[39]][[4]]
#> [1] 0.2
#> 
#> [[39]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[40]]
#> [[40]][[1]]
#> [1] 5.1
#> 
#> [[40]][[2]]
#> [1] 3.4
#> 
#> [[40]][[3]]
#> [1] 1.5
#> 
#> [[40]][[4]]
#> [1] 0.2
#> 
#> [[40]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[41]]
#> [[41]][[1]]
#> [1] 5
#> 
#> [[41]][[2]]
#> [1] 3.5
#> 
#> [[41]][[3]]
#> [1] 1.3
#> 
#> [[41]][[4]]
#> [1] 0.3
#> 
#> [[41]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[42]]
#> [[42]][[1]]
#> [1] 4.5
#> 
#> [[42]][[2]]
#> [1] 2.3
#> 
#> [[42]][[3]]
#> [1] 1.3
#> 
#> [[42]][[4]]
#> [1] 0.3
#> 
#> [[42]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[43]]
#> [[43]][[1]]
#> [1] 4.4
#> 
#> [[43]][[2]]
#> [1] 3.2
#> 
#> [[43]][[3]]
#> [1] 1.3
#> 
#> [[43]][[4]]
#> [1] 0.2
#> 
#> [[43]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[44]]
#> [[44]][[1]]
#> [1] 5
#> 
#> [[44]][[2]]
#> [1] 3.5
#> 
#> [[44]][[3]]
#> [1] 1.6
#> 
#> [[44]][[4]]
#> [1] 0.6
#> 
#> [[44]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[45]]
#> [[45]][[1]]
#> [1] 5.1
#> 
#> [[45]][[2]]
#> [1] 3.8
#> 
#> [[45]][[3]]
#> [1] 1.9
#> 
#> [[45]][[4]]
#> [1] 0.4
#> 
#> [[45]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[46]]
#> [[46]][[1]]
#> [1] 4.8
#> 
#> [[46]][[2]]
#> [1] 3
#> 
#> [[46]][[3]]
#> [1] 1.4
#> 
#> [[46]][[4]]
#> [1] 0.3
#> 
#> [[46]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[47]]
#> [[47]][[1]]
#> [1] 5.1
#> 
#> [[47]][[2]]
#> [1] 3.8
#> 
#> [[47]][[3]]
#> [1] 1.6
#> 
#> [[47]][[4]]
#> [1] 0.2
#> 
#> [[47]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[48]]
#> [[48]][[1]]
#> [1] 4.6
#> 
#> [[48]][[2]]
#> [1] 3.2
#> 
#> [[48]][[3]]
#> [1] 1.4
#> 
#> [[48]][[4]]
#> [1] 0.2
#> 
#> [[48]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[49]]
#> [[49]][[1]]
#> [1] 5.3
#> 
#> [[49]][[2]]
#> [1] 3.7
#> 
#> [[49]][[3]]
#> [1] 1.5
#> 
#> [[49]][[4]]
#> [1] 0.2
#> 
#> [[49]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[50]]
#> [[50]][[1]]
#> [1] 5
#> 
#> [[50]][[2]]
#> [1] 3.3
#> 
#> [[50]][[3]]
#> [1] 1.4
#> 
#> [[50]][[4]]
#> [1] 0.2
#> 
#> [[50]][[5]]
#> [1] "setosa"
#> 
#> 
#> [[51]]
#> [[51]][[1]]
#> [1] 7
#> 
#> [[51]][[2]]
#> [1] 3.2
#> 
#> [[51]][[3]]
#> [1] 4.7
#> 
#> [[51]][[4]]
#> [1] 1.4
#> 
#> [[51]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[52]]
#> [[52]][[1]]
#> [1] 6.4
#> 
#> [[52]][[2]]
#> [1] 3.2
#> 
#> [[52]][[3]]
#> [1] 4.5
#> 
#> [[52]][[4]]
#> [1] 1.5
#> 
#> [[52]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[53]]
#> [[53]][[1]]
#> [1] 6.9
#> 
#> [[53]][[2]]
#> [1] 3.1
#> 
#> [[53]][[3]]
#> [1] 4.9
#> 
#> [[53]][[4]]
#> [1] 1.5
#> 
#> [[53]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[54]]
#> [[54]][[1]]
#> [1] 5.5
#> 
#> [[54]][[2]]
#> [1] 2.3
#> 
#> [[54]][[3]]
#> [1] 4
#> 
#> [[54]][[4]]
#> [1] 1.3
#> 
#> [[54]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[55]]
#> [[55]][[1]]
#> [1] 6.5
#> 
#> [[55]][[2]]
#> [1] 2.8
#> 
#> [[55]][[3]]
#> [1] 4.6
#> 
#> [[55]][[4]]
#> [1] 1.5
#> 
#> [[55]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[56]]
#> [[56]][[1]]
#> [1] 5.7
#> 
#> [[56]][[2]]
#> [1] 2.8
#> 
#> [[56]][[3]]
#> [1] 4.5
#> 
#> [[56]][[4]]
#> [1] 1.3
#> 
#> [[56]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[57]]
#> [[57]][[1]]
#> [1] 6.3
#> 
#> [[57]][[2]]
#> [1] 3.3
#> 
#> [[57]][[3]]
#> [1] 4.7
#> 
#> [[57]][[4]]
#> [1] 1.6
#> 
#> [[57]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[58]]
#> [[58]][[1]]
#> [1] 4.9
#> 
#> [[58]][[2]]
#> [1] 2.4
#> 
#> [[58]][[3]]
#> [1] 3.3
#> 
#> [[58]][[4]]
#> [1] 1
#> 
#> [[58]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[59]]
#> [[59]][[1]]
#> [1] 6.6
#> 
#> [[59]][[2]]
#> [1] 2.9
#> 
#> [[59]][[3]]
#> [1] 4.6
#> 
#> [[59]][[4]]
#> [1] 1.3
#> 
#> [[59]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[60]]
#> [[60]][[1]]
#> [1] 5.2
#> 
#> [[60]][[2]]
#> [1] 2.7
#> 
#> [[60]][[3]]
#> [1] 3.9
#> 
#> [[60]][[4]]
#> [1] 1.4
#> 
#> [[60]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[61]]
#> [[61]][[1]]
#> [1] 5
#> 
#> [[61]][[2]]
#> [1] 2
#> 
#> [[61]][[3]]
#> [1] 3.5
#> 
#> [[61]][[4]]
#> [1] 1
#> 
#> [[61]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[62]]
#> [[62]][[1]]
#> [1] 5.9
#> 
#> [[62]][[2]]
#> [1] 3
#> 
#> [[62]][[3]]
#> [1] 4.2
#> 
#> [[62]][[4]]
#> [1] 1.5
#> 
#> [[62]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[63]]
#> [[63]][[1]]
#> [1] 6
#> 
#> [[63]][[2]]
#> [1] 2.2
#> 
#> [[63]][[3]]
#> [1] 4
#> 
#> [[63]][[4]]
#> [1] 1
#> 
#> [[63]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[64]]
#> [[64]][[1]]
#> [1] 6.1
#> 
#> [[64]][[2]]
#> [1] 2.9
#> 
#> [[64]][[3]]
#> [1] 4.7
#> 
#> [[64]][[4]]
#> [1] 1.4
#> 
#> [[64]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[65]]
#> [[65]][[1]]
#> [1] 5.6
#> 
#> [[65]][[2]]
#> [1] 2.9
#> 
#> [[65]][[3]]
#> [1] 3.6
#> 
#> [[65]][[4]]
#> [1] 1.3
#> 
#> [[65]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[66]]
#> [[66]][[1]]
#> [1] 6.7
#> 
#> [[66]][[2]]
#> [1] 3.1
#> 
#> [[66]][[3]]
#> [1] 4.4
#> 
#> [[66]][[4]]
#> [1] 1.4
#> 
#> [[66]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[67]]
#> [[67]][[1]]
#> [1] 5.6
#> 
#> [[67]][[2]]
#> [1] 3
#> 
#> [[67]][[3]]
#> [1] 4.5
#> 
#> [[67]][[4]]
#> [1] 1.5
#> 
#> [[67]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[68]]
#> [[68]][[1]]
#> [1] 5.8
#> 
#> [[68]][[2]]
#> [1] 2.7
#> 
#> [[68]][[3]]
#> [1] 4.1
#> 
#> [[68]][[4]]
#> [1] 1
#> 
#> [[68]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[69]]
#> [[69]][[1]]
#> [1] 6.2
#> 
#> [[69]][[2]]
#> [1] 2.2
#> 
#> [[69]][[3]]
#> [1] 4.5
#> 
#> [[69]][[4]]
#> [1] 1.5
#> 
#> [[69]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[70]]
#> [[70]][[1]]
#> [1] 5.6
#> 
#> [[70]][[2]]
#> [1] 2.5
#> 
#> [[70]][[3]]
#> [1] 3.9
#> 
#> [[70]][[4]]
#> [1] 1.1
#> 
#> [[70]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[71]]
#> [[71]][[1]]
#> [1] 5.9
#> 
#> [[71]][[2]]
#> [1] 3.2
#> 
#> [[71]][[3]]
#> [1] 4.8
#> 
#> [[71]][[4]]
#> [1] 1.8
#> 
#> [[71]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[72]]
#> [[72]][[1]]
#> [1] 6.1
#> 
#> [[72]][[2]]
#> [1] 2.8
#> 
#> [[72]][[3]]
#> [1] 4
#> 
#> [[72]][[4]]
#> [1] 1.3
#> 
#> [[72]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[73]]
#> [[73]][[1]]
#> [1] 6.3
#> 
#> [[73]][[2]]
#> [1] 2.5
#> 
#> [[73]][[3]]
#> [1] 4.9
#> 
#> [[73]][[4]]
#> [1] 1.5
#> 
#> [[73]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[74]]
#> [[74]][[1]]
#> [1] 6.1
#> 
#> [[74]][[2]]
#> [1] 2.8
#> 
#> [[74]][[3]]
#> [1] 4.7
#> 
#> [[74]][[4]]
#> [1] 1.2
#> 
#> [[74]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[75]]
#> [[75]][[1]]
#> [1] 6.4
#> 
#> [[75]][[2]]
#> [1] 2.9
#> 
#> [[75]][[3]]
#> [1] 4.3
#> 
#> [[75]][[4]]
#> [1] 1.3
#> 
#> [[75]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[76]]
#> [[76]][[1]]
#> [1] 6.6
#> 
#> [[76]][[2]]
#> [1] 3
#> 
#> [[76]][[3]]
#> [1] 4.4
#> 
#> [[76]][[4]]
#> [1] 1.4
#> 
#> [[76]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[77]]
#> [[77]][[1]]
#> [1] 6.8
#> 
#> [[77]][[2]]
#> [1] 2.8
#> 
#> [[77]][[3]]
#> [1] 4.8
#> 
#> [[77]][[4]]
#> [1] 1.4
#> 
#> [[77]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[78]]
#> [[78]][[1]]
#> [1] 6.7
#> 
#> [[78]][[2]]
#> [1] 3
#> 
#> [[78]][[3]]
#> [1] 5
#> 
#> [[78]][[4]]
#> [1] 1.7
#> 
#> [[78]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[79]]
#> [[79]][[1]]
#> [1] 6
#> 
#> [[79]][[2]]
#> [1] 2.9
#> 
#> [[79]][[3]]
#> [1] 4.5
#> 
#> [[79]][[4]]
#> [1] 1.5
#> 
#> [[79]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[80]]
#> [[80]][[1]]
#> [1] 5.7
#> 
#> [[80]][[2]]
#> [1] 2.6
#> 
#> [[80]][[3]]
#> [1] 3.5
#> 
#> [[80]][[4]]
#> [1] 1
#> 
#> [[80]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[81]]
#> [[81]][[1]]
#> [1] 5.5
#> 
#> [[81]][[2]]
#> [1] 2.4
#> 
#> [[81]][[3]]
#> [1] 3.8
#> 
#> [[81]][[4]]
#> [1] 1.1
#> 
#> [[81]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[82]]
#> [[82]][[1]]
#> [1] 5.5
#> 
#> [[82]][[2]]
#> [1] 2.4
#> 
#> [[82]][[3]]
#> [1] 3.7
#> 
#> [[82]][[4]]
#> [1] 1
#> 
#> [[82]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[83]]
#> [[83]][[1]]
#> [1] 5.8
#> 
#> [[83]][[2]]
#> [1] 2.7
#> 
#> [[83]][[3]]
#> [1] 3.9
#> 
#> [[83]][[4]]
#> [1] 1.2
#> 
#> [[83]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[84]]
#> [[84]][[1]]
#> [1] 6
#> 
#> [[84]][[2]]
#> [1] 2.7
#> 
#> [[84]][[3]]
#> [1] 5.1
#> 
#> [[84]][[4]]
#> [1] 1.6
#> 
#> [[84]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[85]]
#> [[85]][[1]]
#> [1] 5.4
#> 
#> [[85]][[2]]
#> [1] 3
#> 
#> [[85]][[3]]
#> [1] 4.5
#> 
#> [[85]][[4]]
#> [1] 1.5
#> 
#> [[85]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[86]]
#> [[86]][[1]]
#> [1] 6
#> 
#> [[86]][[2]]
#> [1] 3.4
#> 
#> [[86]][[3]]
#> [1] 4.5
#> 
#> [[86]][[4]]
#> [1] 1.6
#> 
#> [[86]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[87]]
#> [[87]][[1]]
#> [1] 6.7
#> 
#> [[87]][[2]]
#> [1] 3.1
#> 
#> [[87]][[3]]
#> [1] 4.7
#> 
#> [[87]][[4]]
#> [1] 1.5
#> 
#> [[87]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[88]]
#> [[88]][[1]]
#> [1] 6.3
#> 
#> [[88]][[2]]
#> [1] 2.3
#> 
#> [[88]][[3]]
#> [1] 4.4
#> 
#> [[88]][[4]]
#> [1] 1.3
#> 
#> [[88]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[89]]
#> [[89]][[1]]
#> [1] 5.6
#> 
#> [[89]][[2]]
#> [1] 3
#> 
#> [[89]][[3]]
#> [1] 4.1
#> 
#> [[89]][[4]]
#> [1] 1.3
#> 
#> [[89]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[90]]
#> [[90]][[1]]
#> [1] 5.5
#> 
#> [[90]][[2]]
#> [1] 2.5
#> 
#> [[90]][[3]]
#> [1] 4
#> 
#> [[90]][[4]]
#> [1] 1.3
#> 
#> [[90]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[91]]
#> [[91]][[1]]
#> [1] 5.5
#> 
#> [[91]][[2]]
#> [1] 2.6
#> 
#> [[91]][[3]]
#> [1] 4.4
#> 
#> [[91]][[4]]
#> [1] 1.2
#> 
#> [[91]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[92]]
#> [[92]][[1]]
#> [1] 6.1
#> 
#> [[92]][[2]]
#> [1] 3
#> 
#> [[92]][[3]]
#> [1] 4.6
#> 
#> [[92]][[4]]
#> [1] 1.4
#> 
#> [[92]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[93]]
#> [[93]][[1]]
#> [1] 5.8
#> 
#> [[93]][[2]]
#> [1] 2.6
#> 
#> [[93]][[3]]
#> [1] 4
#> 
#> [[93]][[4]]
#> [1] 1.2
#> 
#> [[93]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[94]]
#> [[94]][[1]]
#> [1] 5
#> 
#> [[94]][[2]]
#> [1] 2.3
#> 
#> [[94]][[3]]
#> [1] 3.3
#> 
#> [[94]][[4]]
#> [1] 1
#> 
#> [[94]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[95]]
#> [[95]][[1]]
#> [1] 5.6
#> 
#> [[95]][[2]]
#> [1] 2.7
#> 
#> [[95]][[3]]
#> [1] 4.2
#> 
#> [[95]][[4]]
#> [1] 1.3
#> 
#> [[95]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[96]]
#> [[96]][[1]]
#> [1] 5.7
#> 
#> [[96]][[2]]
#> [1] 3
#> 
#> [[96]][[3]]
#> [1] 4.2
#> 
#> [[96]][[4]]
#> [1] 1.2
#> 
#> [[96]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[97]]
#> [[97]][[1]]
#> [1] 5.7
#> 
#> [[97]][[2]]
#> [1] 2.9
#> 
#> [[97]][[3]]
#> [1] 4.2
#> 
#> [[97]][[4]]
#> [1] 1.3
#> 
#> [[97]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[98]]
#> [[98]][[1]]
#> [1] 6.2
#> 
#> [[98]][[2]]
#> [1] 2.9
#> 
#> [[98]][[3]]
#> [1] 4.3
#> 
#> [[98]][[4]]
#> [1] 1.3
#> 
#> [[98]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[99]]
#> [[99]][[1]]
#> [1] 5.1
#> 
#> [[99]][[2]]
#> [1] 2.5
#> 
#> [[99]][[3]]
#> [1] 3
#> 
#> [[99]][[4]]
#> [1] 1.1
#> 
#> [[99]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[100]]
#> [[100]][[1]]
#> [1] 5.7
#> 
#> [[100]][[2]]
#> [1] 2.8
#> 
#> [[100]][[3]]
#> [1] 4.1
#> 
#> [[100]][[4]]
#> [1] 1.3
#> 
#> [[100]][[5]]
#> [1] "versicolor"
#> 
#> 
#> [[101]]
#> [[101]][[1]]
#> [1] 6.3
#> 
#> [[101]][[2]]
#> [1] 3.3
#> 
#> [[101]][[3]]
#> [1] 6
#> 
#> [[101]][[4]]
#> [1] 2.5
#> 
#> [[101]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[102]]
#> [[102]][[1]]
#> [1] 5.8
#> 
#> [[102]][[2]]
#> [1] 2.7
#> 
#> [[102]][[3]]
#> [1] 5.1
#> 
#> [[102]][[4]]
#> [1] 1.9
#> 
#> [[102]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[103]]
#> [[103]][[1]]
#> [1] 7.1
#> 
#> [[103]][[2]]
#> [1] 3
#> 
#> [[103]][[3]]
#> [1] 5.9
#> 
#> [[103]][[4]]
#> [1] 2.1
#> 
#> [[103]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[104]]
#> [[104]][[1]]
#> [1] 6.3
#> 
#> [[104]][[2]]
#> [1] 2.9
#> 
#> [[104]][[3]]
#> [1] 5.6
#> 
#> [[104]][[4]]
#> [1] 1.8
#> 
#> [[104]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[105]]
#> [[105]][[1]]
#> [1] 6.5
#> 
#> [[105]][[2]]
#> [1] 3
#> 
#> [[105]][[3]]
#> [1] 5.8
#> 
#> [[105]][[4]]
#> [1] 2.2
#> 
#> [[105]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[106]]
#> [[106]][[1]]
#> [1] 7.6
#> 
#> [[106]][[2]]
#> [1] 3
#> 
#> [[106]][[3]]
#> [1] 6.6
#> 
#> [[106]][[4]]
#> [1] 2.1
#> 
#> [[106]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[107]]
#> [[107]][[1]]
#> [1] 4.9
#> 
#> [[107]][[2]]
#> [1] 2.5
#> 
#> [[107]][[3]]
#> [1] 4.5
#> 
#> [[107]][[4]]
#> [1] 1.7
#> 
#> [[107]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[108]]
#> [[108]][[1]]
#> [1] 7.3
#> 
#> [[108]][[2]]
#> [1] 2.9
#> 
#> [[108]][[3]]
#> [1] 6.3
#> 
#> [[108]][[4]]
#> [1] 1.8
#> 
#> [[108]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[109]]
#> [[109]][[1]]
#> [1] 6.7
#> 
#> [[109]][[2]]
#> [1] 2.5
#> 
#> [[109]][[3]]
#> [1] 5.8
#> 
#> [[109]][[4]]
#> [1] 1.8
#> 
#> [[109]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[110]]
#> [[110]][[1]]
#> [1] 7.2
#> 
#> [[110]][[2]]
#> [1] 3.6
#> 
#> [[110]][[3]]
#> [1] 6.1
#> 
#> [[110]][[4]]
#> [1] 2.5
#> 
#> [[110]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[111]]
#> [[111]][[1]]
#> [1] 6.5
#> 
#> [[111]][[2]]
#> [1] 3.2
#> 
#> [[111]][[3]]
#> [1] 5.1
#> 
#> [[111]][[4]]
#> [1] 2
#> 
#> [[111]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[112]]
#> [[112]][[1]]
#> [1] 6.4
#> 
#> [[112]][[2]]
#> [1] 2.7
#> 
#> [[112]][[3]]
#> [1] 5.3
#> 
#> [[112]][[4]]
#> [1] 1.9
#> 
#> [[112]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[113]]
#> [[113]][[1]]
#> [1] 6.8
#> 
#> [[113]][[2]]
#> [1] 3
#> 
#> [[113]][[3]]
#> [1] 5.5
#> 
#> [[113]][[4]]
#> [1] 2.1
#> 
#> [[113]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[114]]
#> [[114]][[1]]
#> [1] 5.7
#> 
#> [[114]][[2]]
#> [1] 2.5
#> 
#> [[114]][[3]]
#> [1] 5
#> 
#> [[114]][[4]]
#> [1] 2
#> 
#> [[114]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[115]]
#> [[115]][[1]]
#> [1] 5.8
#> 
#> [[115]][[2]]
#> [1] 2.8
#> 
#> [[115]][[3]]
#> [1] 5.1
#> 
#> [[115]][[4]]
#> [1] 2.4
#> 
#> [[115]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[116]]
#> [[116]][[1]]
#> [1] 6.4
#> 
#> [[116]][[2]]
#> [1] 3.2
#> 
#> [[116]][[3]]
#> [1] 5.3
#> 
#> [[116]][[4]]
#> [1] 2.3
#> 
#> [[116]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[117]]
#> [[117]][[1]]
#> [1] 6.5
#> 
#> [[117]][[2]]
#> [1] 3
#> 
#> [[117]][[3]]
#> [1] 5.5
#> 
#> [[117]][[4]]
#> [1] 1.8
#> 
#> [[117]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[118]]
#> [[118]][[1]]
#> [1] 7.7
#> 
#> [[118]][[2]]
#> [1] 3.8
#> 
#> [[118]][[3]]
#> [1] 6.7
#> 
#> [[118]][[4]]
#> [1] 2.2
#> 
#> [[118]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[119]]
#> [[119]][[1]]
#> [1] 7.7
#> 
#> [[119]][[2]]
#> [1] 2.6
#> 
#> [[119]][[3]]
#> [1] 6.9
#> 
#> [[119]][[4]]
#> [1] 2.3
#> 
#> [[119]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[120]]
#> [[120]][[1]]
#> [1] 6
#> 
#> [[120]][[2]]
#> [1] 2.2
#> 
#> [[120]][[3]]
#> [1] 5
#> 
#> [[120]][[4]]
#> [1] 1.5
#> 
#> [[120]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[121]]
#> [[121]][[1]]
#> [1] 6.9
#> 
#> [[121]][[2]]
#> [1] 3.2
#> 
#> [[121]][[3]]
#> [1] 5.7
#> 
#> [[121]][[4]]
#> [1] 2.3
#> 
#> [[121]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[122]]
#> [[122]][[1]]
#> [1] 5.6
#> 
#> [[122]][[2]]
#> [1] 2.8
#> 
#> [[122]][[3]]
#> [1] 4.9
#> 
#> [[122]][[4]]
#> [1] 2
#> 
#> [[122]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[123]]
#> [[123]][[1]]
#> [1] 7.7
#> 
#> [[123]][[2]]
#> [1] 2.8
#> 
#> [[123]][[3]]
#> [1] 6.7
#> 
#> [[123]][[4]]
#> [1] 2
#> 
#> [[123]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[124]]
#> [[124]][[1]]
#> [1] 6.3
#> 
#> [[124]][[2]]
#> [1] 2.7
#> 
#> [[124]][[3]]
#> [1] 4.9
#> 
#> [[124]][[4]]
#> [1] 1.8
#> 
#> [[124]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[125]]
#> [[125]][[1]]
#> [1] 6.7
#> 
#> [[125]][[2]]
#> [1] 3.3
#> 
#> [[125]][[3]]
#> [1] 5.7
#> 
#> [[125]][[4]]
#> [1] 2.1
#> 
#> [[125]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[126]]
#> [[126]][[1]]
#> [1] 7.2
#> 
#> [[126]][[2]]
#> [1] 3.2
#> 
#> [[126]][[3]]
#> [1] 6
#> 
#> [[126]][[4]]
#> [1] 1.8
#> 
#> [[126]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[127]]
#> [[127]][[1]]
#> [1] 6.2
#> 
#> [[127]][[2]]
#> [1] 2.8
#> 
#> [[127]][[3]]
#> [1] 4.8
#> 
#> [[127]][[4]]
#> [1] 1.8
#> 
#> [[127]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[128]]
#> [[128]][[1]]
#> [1] 6.1
#> 
#> [[128]][[2]]
#> [1] 3
#> 
#> [[128]][[3]]
#> [1] 4.9
#> 
#> [[128]][[4]]
#> [1] 1.8
#> 
#> [[128]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[129]]
#> [[129]][[1]]
#> [1] 6.4
#> 
#> [[129]][[2]]
#> [1] 2.8
#> 
#> [[129]][[3]]
#> [1] 5.6
#> 
#> [[129]][[4]]
#> [1] 2.1
#> 
#> [[129]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[130]]
#> [[130]][[1]]
#> [1] 7.2
#> 
#> [[130]][[2]]
#> [1] 3
#> 
#> [[130]][[3]]
#> [1] 5.8
#> 
#> [[130]][[4]]
#> [1] 1.6
#> 
#> [[130]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[131]]
#> [[131]][[1]]
#> [1] 7.4
#> 
#> [[131]][[2]]
#> [1] 2.8
#> 
#> [[131]][[3]]
#> [1] 6.1
#> 
#> [[131]][[4]]
#> [1] 1.9
#> 
#> [[131]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[132]]
#> [[132]][[1]]
#> [1] 7.9
#> 
#> [[132]][[2]]
#> [1] 3.8
#> 
#> [[132]][[3]]
#> [1] 6.4
#> 
#> [[132]][[4]]
#> [1] 2
#> 
#> [[132]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[133]]
#> [[133]][[1]]
#> [1] 6.4
#> 
#> [[133]][[2]]
#> [1] 2.8
#> 
#> [[133]][[3]]
#> [1] 5.6
#> 
#> [[133]][[4]]
#> [1] 2.2
#> 
#> [[133]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[134]]
#> [[134]][[1]]
#> [1] 6.3
#> 
#> [[134]][[2]]
#> [1] 2.8
#> 
#> [[134]][[3]]
#> [1] 5.1
#> 
#> [[134]][[4]]
#> [1] 1.5
#> 
#> [[134]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[135]]
#> [[135]][[1]]
#> [1] 6.1
#> 
#> [[135]][[2]]
#> [1] 2.6
#> 
#> [[135]][[3]]
#> [1] 5.6
#> 
#> [[135]][[4]]
#> [1] 1.4
#> 
#> [[135]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[136]]
#> [[136]][[1]]
#> [1] 7.7
#> 
#> [[136]][[2]]
#> [1] 3
#> 
#> [[136]][[3]]
#> [1] 6.1
#> 
#> [[136]][[4]]
#> [1] 2.3
#> 
#> [[136]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[137]]
#> [[137]][[1]]
#> [1] 6.3
#> 
#> [[137]][[2]]
#> [1] 3.4
#> 
#> [[137]][[3]]
#> [1] 5.6
#> 
#> [[137]][[4]]
#> [1] 2.4
#> 
#> [[137]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[138]]
#> [[138]][[1]]
#> [1] 6.4
#> 
#> [[138]][[2]]
#> [1] 3.1
#> 
#> [[138]][[3]]
#> [1] 5.5
#> 
#> [[138]][[4]]
#> [1] 1.8
#> 
#> [[138]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[139]]
#> [[139]][[1]]
#> [1] 6
#> 
#> [[139]][[2]]
#> [1] 3
#> 
#> [[139]][[3]]
#> [1] 4.8
#> 
#> [[139]][[4]]
#> [1] 1.8
#> 
#> [[139]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[140]]
#> [[140]][[1]]
#> [1] 6.9
#> 
#> [[140]][[2]]
#> [1] 3.1
#> 
#> [[140]][[3]]
#> [1] 5.4
#> 
#> [[140]][[4]]
#> [1] 2.1
#> 
#> [[140]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[141]]
#> [[141]][[1]]
#> [1] 6.7
#> 
#> [[141]][[2]]
#> [1] 3.1
#> 
#> [[141]][[3]]
#> [1] 5.6
#> 
#> [[141]][[4]]
#> [1] 2.4
#> 
#> [[141]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[142]]
#> [[142]][[1]]
#> [1] 6.9
#> 
#> [[142]][[2]]
#> [1] 3.1
#> 
#> [[142]][[3]]
#> [1] 5.1
#> 
#> [[142]][[4]]
#> [1] 2.3
#> 
#> [[142]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[143]]
#> [[143]][[1]]
#> [1] 5.8
#> 
#> [[143]][[2]]
#> [1] 2.7
#> 
#> [[143]][[3]]
#> [1] 5.1
#> 
#> [[143]][[4]]
#> [1] 1.9
#> 
#> [[143]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[144]]
#> [[144]][[1]]
#> [1] 6.8
#> 
#> [[144]][[2]]
#> [1] 3.2
#> 
#> [[144]][[3]]
#> [1] 5.9
#> 
#> [[144]][[4]]
#> [1] 2.3
#> 
#> [[144]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[145]]
#> [[145]][[1]]
#> [1] 6.7
#> 
#> [[145]][[2]]
#> [1] 3.3
#> 
#> [[145]][[3]]
#> [1] 5.7
#> 
#> [[145]][[4]]
#> [1] 2.5
#> 
#> [[145]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[146]]
#> [[146]][[1]]
#> [1] 6.7
#> 
#> [[146]][[2]]
#> [1] 3
#> 
#> [[146]][[3]]
#> [1] 5.2
#> 
#> [[146]][[4]]
#> [1] 2.3
#> 
#> [[146]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[147]]
#> [[147]][[1]]
#> [1] 6.3
#> 
#> [[147]][[2]]
#> [1] 2.5
#> 
#> [[147]][[3]]
#> [1] 5
#> 
#> [[147]][[4]]
#> [1] 1.9
#> 
#> [[147]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[148]]
#> [[148]][[1]]
#> [1] 6.5
#> 
#> [[148]][[2]]
#> [1] 3
#> 
#> [[148]][[3]]
#> [1] 5.2
#> 
#> [[148]][[4]]
#> [1] 2
#> 
#> [[148]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[149]]
#> [[149]][[1]]
#> [1] 6.2
#> 
#> [[149]][[2]]
#> [1] 3.4
#> 
#> [[149]][[3]]
#> [1] 5.4
#> 
#> [[149]][[4]]
#> [1] 2.3
#> 
#> [[149]][[5]]
#> [1] "virginica"
#> 
#> 
#> [[150]]
#> [[150]][[1]]
#> [1] 5.9
#> 
#> [[150]][[2]]
#> [1] 3
#> 
#> [[150]][[3]]
#> [1] 5.1
#> 
#> [[150]][[4]]
#> [1] 1.8
#> 
#> [[150]][[5]]
#> [1] "virginica"
#> 
#> 

# Keep variables names
parse_df(iris[, 1:2], add_names = TRUE)

#> [[1]]
#> [[1]]$Sepal.Length
#> [1] 5.1
#> 
#> [[1]]$Sepal.Width
#> [1] 3.5
#> 
#> 
#> [[2]]
#> [[2]]$Sepal.Length
#> [1] 4.9
#> 
#> [[2]]$Sepal.Width
#> [1] 3
#> 
#> 
#> [[3]]
#> [[3]]$Sepal.Length
#> [1] 4.7
#> 
#> [[3]]$Sepal.Width
#> [1] 3.2
#> 
#> 
#> [[4]]
#> [[4]]$Sepal.Length
#> [1] 4.6
#> 
#> [[4]]$Sepal.Width
#> [1] 3.1
#> 
#> 
#> [[5]]
#> [[5]]$Sepal.Length
#> [1] 5
#> 
#> [[5]]$Sepal.Width
#> [1] 3.6
#> 
#> 
#> [[6]]
#> [[6]]$Sepal.Length
#> [1] 5.4
#> 
#> [[6]]$Sepal.Width
#> [1] 3.9
#> 
#> 
#> [[7]]
#> [[7]]$Sepal.Length
#> [1] 4.6
#> 
#> [[7]]$Sepal.Width
#> [1] 3.4
#> 
#> 
#> [[8]]
#> [[8]]$Sepal.Length
#> [1] 5
#> 
#> [[8]]$Sepal.Width
#> [1] 3.4
#> 
#> 
#> [[9]]
#> [[9]]$Sepal.Length
#> [1] 4.4
#> 
#> [[9]]$Sepal.Width
#> [1] 2.9
#> 
#> 
#> [[10]]
#> [[10]]$Sepal.Length
#> [1] 4.9
#> 
#> [[10]]$Sepal.Width
#> [1] 3.1
#> 
#> 
#> [[11]]
#> [[11]]$Sepal.Length
#> [1] 5.4
#> 
#> [[11]]$Sepal.Width
#> [1] 3.7
#> 
#> 
#> [[12]]
#> [[12]]$Sepal.Length
#> [1] 4.8
#> 
#> [[12]]$Sepal.Width
#> [1] 3.4
#> 
#> 
#> [[13]]
#> [[13]]$Sepal.Length
#> [1] 4.8
#> 
#> [[13]]$Sepal.Width
#> [1] 3
#> 
#> 
#> [[14]]
#> [[14]]$Sepal.Length
#> [1] 4.3
#> 
#> [[14]]$Sepal.Width
#> [1] 3
#> 
#> 
#> [[15]]
#> [[15]]$Sepal.Length
#> [1] 5.8
#> 
#> [[15]]$Sepal.Width
#> [1] 4
#> 
#> 
#> [[16]]
#> [[16]]$Sepal.Length
#> [1] 5.7
#> 
#> [[16]]$Sepal.Width
#> [1] 4.4
#> 
#> 
#> [[17]]
#> [[17]]$Sepal.Length
#> [1] 5.4
#> 
#> [[17]]$Sepal.Width
#> [1] 3.9
#> 
#> 
#> [[18]]
#> [[18]]$Sepal.Length
#> [1] 5.1
#> 
#> [[18]]$Sepal.Width
#> [1] 3.5
#> 
#> 
#> [[19]]
#> [[19]]$Sepal.Length
#> [1] 5.7
#> 
#> [[19]]$Sepal.Width
#> [1] 3.8
#> 
#> 
#> [[20]]
#> [[20]]$Sepal.Length
#> [1] 5.1
#> 
#> [[20]]$Sepal.Width
#> [1] 3.8
#> 
#> 
#> [[21]]
#> [[21]]$Sepal.Length
#> [1] 5.4
#> 
#> [[21]]$Sepal.Width
#> [1] 3.4
#> 
#> 
#> [[22]]
#> [[22]]$Sepal.Length
#> [1] 5.1
#> 
#> [[22]]$Sepal.Width
#> [1] 3.7
#> 
#> 
#> [[23]]
#> [[23]]$Sepal.Length
#> [1] 4.6
#> 
#> [[23]]$Sepal.Width
#> [1] 3.6
#> 
#> 
#> [[24]]
#> [[24]]$Sepal.Length
#> [1] 5.1
#> 
#> [[24]]$Sepal.Width
#> [1] 3.3
#> 
#> 
#> [[25]]
#> [[25]]$Sepal.Length
#> [1] 4.8
#> 
#> [[25]]$Sepal.Width
#> [1] 3.4
#> 
#> 
#> [[26]]
#> [[26]]$Sepal.Length
#> [1] 5
#> 
#> [[26]]$Sepal.Width
#> [1] 3
#> 
#> 
#> [[27]]
#> [[27]]$Sepal.Length
#> [1] 5
#> 
#> [[27]]$Sepal.Width
#> [1] 3.4
#> 
#> 
#> [[28]]
#> [[28]]$Sepal.Length
#> [1] 5.2
#> 
#> [[28]]$Sepal.Width
#> [1] 3.5
#> 
#> 
#> [[29]]
#> [[29]]$Sepal.Length
#> [1] 5.2
#> 
#> [[29]]$Sepal.Width
#> [1] 3.4
#> 
#> 
#> [[30]]
#> [[30]]$Sepal.Length
#> [1] 4.7
#> 
#> [[30]]$Sepal.Width
#> [1] 3.2
#> 
#> 
#> [[31]]
#> [[31]]$Sepal.Length
#> [1] 4.8
#> 
#> [[31]]$Sepal.Width
#> [1] 3.1
#> 
#> 
#> [[32]]
#> [[32]]$Sepal.Length
#> [1] 5.4
#> 
#> [[32]]$Sepal.Width
#> [1] 3.4
#> 
#> 
#> [[33]]
#> [[33]]$Sepal.Length
#> [1] 5.2
#> 
#> [[33]]$Sepal.Width
#> [1] 4.1
#> 
#> 
#> [[34]]
#> [[34]]$Sepal.Length
#> [1] 5.5
#> 
#> [[34]]$Sepal.Width
#> [1] 4.2
#> 
#> 
#> [[35]]
#> [[35]]$Sepal.Length
#> [1] 4.9
#> 
#> [[35]]$Sepal.Width
#> [1] 3.1
#> 
#> 
#> [[36]]
#> [[36]]$Sepal.Length
#> [1] 5
#> 
#> [[36]]$Sepal.Width
#> [1] 3.2
#> 
#> 
#> [[37]]
#> [[37]]$Sepal.Length
#> [1] 5.5
#> 
#> [[37]]$Sepal.Width
#> [1] 3.5
#> 
#> 
#> [[38]]
#> [[38]]$Sepal.Length
#> [1] 4.9
#> 
#> [[38]]$Sepal.Width
#> [1] 3.6
#> 
#> 
#> [[39]]
#> [[39]]$Sepal.Length
#> [1] 4.4
#> 
#> [[39]]$Sepal.Width
#> [1] 3
#> 
#> 
#> [[40]]
#> [[40]]$Sepal.Length
#> [1] 5.1
#> 
#> [[40]]$Sepal.Width
#> [1] 3.4
#> 
#> 
#> [[41]]
#> [[41]]$Sepal.Length
#> [1] 5
#> 
#> [[41]]$Sepal.Width
#> [1] 3.5
#> 
#> 
#> [[42]]
#> [[42]]$Sepal.Length
#> [1] 4.5
#> 
#> [[42]]$Sepal.Width
#> [1] 2.3
#> 
#> 
#> [[43]]
#> [[43]]$Sepal.Length
#> [1] 4.4
#> 
#> [[43]]$Sepal.Width
#> [1] 3.2
#> 
#> 
#> [[44]]
#> [[44]]$Sepal.Length
#> [1] 5
#> 
#> [[44]]$Sepal.Width
#> [1] 3.5
#> 
#> 
#> [[45]]
#> [[45]]$Sepal.Length
#> [1] 5.1
#> 
#> [[45]]$Sepal.Width
#> [1] 3.8
#> 
#> 
#> [[46]]
#> [[46]]$Sepal.Length
#> [1] 4.8
#> 
#> [[46]]$Sepal.Width
#> [1] 3
#> 
#> 
#> [[47]]
#> [[47]]$Sepal.Length
#> [1] 5.1
#> 
#> [[47]]$Sepal.Width
#> [1] 3.8
#> 
#> 
#> [[48]]
#> [[48]]$Sepal.Length
#> [1] 4.6
#> 
#> [[48]]$Sepal.Width
#> [1] 3.2
#> 
#> 
#> [[49]]
#> [[49]]$Sepal.Length
#> [1] 5.3
#> 
#> [[49]]$Sepal.Width
#> [1] 3.7
#> 
#> 
#> [[50]]
#> [[50]]$Sepal.Length
#> [1] 5
#> 
#> [[50]]$Sepal.Width
#> [1] 3.3
#> 
#> 
#> [[51]]
#> [[51]]$Sepal.Length
#> [1] 7
#> 
#> [[51]]$Sepal.Width
#> [1] 3.2
#> 
#> 
#> [[52]]
#> [[52]]$Sepal.Length
#> [1] 6.4
#> 
#> [[52]]$Sepal.Width
#> [1] 3.2
#> 
#> 
#> [[53]]
#> [[53]]$Sepal.Length
#> [1] 6.9
#> 
#> [[53]]$Sepal.Width
#> [1] 3.1
#> 
#> 
#> [[54]]
#> [[54]]$Sepal.Length
#> [1] 5.5
#> 
#> [[54]]$Sepal.Width
#> [1] 2.3
#> 
#> 
#> [[55]]
#> [[55]]$Sepal.Length
#> [1] 6.5
#> 
#> [[55]]$Sepal.Width
#> [1] 2.8
#> 
#> 
#> [[56]]
#> [[56]]$Sepal.Length
#> [1] 5.7
#> 
#> [[56]]$Sepal.Width
#> [1] 2.8
#> 
#> 
#> [[57]]
#> [[57]]$Sepal.Length
#> [1] 6.3
#> 
#> [[57]]$Sepal.Width
#> [1] 3.3
#> 
#> 
#> [[58]]
#> [[58]]$Sepal.Length
#> [1] 4.9
#> 
#> [[58]]$Sepal.Width
#> [1] 2.4
#> 
#> 
#> [[59]]
#> [[59]]$Sepal.Length
#> [1] 6.6
#> 
#> [[59]]$Sepal.Width
#> [1] 2.9
#> 
#> 
#> [[60]]
#> [[60]]$Sepal.Length
#> [1] 5.2
#> 
#> [[60]]$Sepal.Width
#> [1] 2.7
#> 
#> 
#> [[61]]
#> [[61]]$Sepal.Length
#> [1] 5
#> 
#> [[61]]$Sepal.Width
#> [1] 2
#> 
#> 
#> [[62]]
#> [[62]]$Sepal.Length
#> [1] 5.9
#> 
#> [[62]]$Sepal.Width
#> [1] 3
#> 
#> 
#> [[63]]
#> [[63]]$Sepal.Length
#> [1] 6
#> 
#> [[63]]$Sepal.Width
#> [1] 2.2
#> 
#> 
#> [[64]]
#> [[64]]$Sepal.Length
#> [1] 6.1
#> 
#> [[64]]$Sepal.Width
#> [1] 2.9
#> 
#> 
#> [[65]]
#> [[65]]$Sepal.Length
#> [1] 5.6
#> 
#> [[65]]$Sepal.Width
#> [1] 2.9
#> 
#> 
#> [[66]]
#> [[66]]$Sepal.Length
#> [1] 6.7
#> 
#> [[66]]$Sepal.Width
#> [1] 3.1
#> 
#> 
#> [[67]]
#> [[67]]$Sepal.Length
#> [1] 5.6
#> 
#> [[67]]$Sepal.Width
#> [1] 3
#> 
#> 
#> [[68]]
#> [[68]]$Sepal.Length
#> [1] 5.8
#> 
#> [[68]]$Sepal.Width
#> [1] 2.7
#> 
#> 
#> [[69]]
#> [[69]]$Sepal.Length
#> [1] 6.2
#> 
#> [[69]]$Sepal.Width
#> [1] 2.2
#> 
#> 
#> [[70]]
#> [[70]]$Sepal.Length
#> [1] 5.6
#> 
#> [[70]]$Sepal.Width
#> [1] 2.5
#> 
#> 
#> [[71]]
#> [[71]]$Sepal.Length
#> [1] 5.9
#> 
#> [[71]]$Sepal.Width
#> [1] 3.2
#> 
#> 
#> [[72]]
#> [[72]]$Sepal.Length
#> [1] 6.1
#> 
#> [[72]]$Sepal.Width
#> [1] 2.8
#> 
#> 
#> [[73]]
#> [[73]]$Sepal.Length
#> [1] 6.3
#> 
#> [[73]]$Sepal.Width
#> [1] 2.5
#> 
#> 
#> [[74]]
#> [[74]]$Sepal.Length
#> [1] 6.1
#> 
#> [[74]]$Sepal.Width
#> [1] 2.8
#> 
#> 
#> [[75]]
#> [[75]]$Sepal.Length
#> [1] 6.4
#> 
#> [[75]]$Sepal.Width
#> [1] 2.9
#> 
#> 
#> [[76]]
#> [[76]]$Sepal.Length
#> [1] 6.6
#> 
#> [[76]]$Sepal.Width
#> [1] 3
#> 
#> 
#> [[77]]
#> [[77]]$Sepal.Length
#> [1] 6.8
#> 
#> [[77]]$Sepal.Width
#> [1] 2.8
#> 
#> 
#> [[78]]
#> [[78]]$Sepal.Length
#> [1] 6.7
#> 
#> [[78]]$Sepal.Width
#> [1] 3
#> 
#> 
#> [[79]]
#> [[79]]$Sepal.Length
#> [1] 6
#> 
#> [[79]]$Sepal.Width
#> [1] 2.9
#> 
#> 
#> [[80]]
#> [[80]]$Sepal.Length
#> [1] 5.7
#> 
#> [[80]]$Sepal.Width
#> [1] 2.6
#> 
#> 
#> [[81]]
#> [[81]]$Sepal.Length
#> [1] 5.5
#> 
#> [[81]]$Sepal.Width
#> [1] 2.4
#> 
#> 
#> [[82]]
#> [[82]]$Sepal.Length
#> [1] 5.5
#> 
#> [[82]]$Sepal.Width
#> [1] 2.4
#> 
#> 
#> [[83]]
#> [[83]]$Sepal.Length
#> [1] 5.8
#> 
#> [[83]]$Sepal.Width
#> [1] 2.7
#> 
#> 
#> [[84]]
#> [[84]]$Sepal.Length
#> [1] 6
#> 
#> [[84]]$Sepal.Width
#> [1] 2.7
#> 
#> 
#> [[85]]
#> [[85]]$Sepal.Length
#> [1] 5.4
#> 
#> [[85]]$Sepal.Width
#> [1] 3
#> 
#> 
#> [[86]]
#> [[86]]$Sepal.Length
#> [1] 6
#> 
#> [[86]]$Sepal.Width
#> [1] 3.4
#> 
#> 
#> [[87]]
#> [[87]]$Sepal.Length
#> [1] 6.7
#> 
#> [[87]]$Sepal.Width
#> [1] 3.1
#> 
#> 
#> [[88]]
#> [[88]]$Sepal.Length
#> [1] 6.3
#> 
#> [[88]]$Sepal.Width
#> [1] 2.3
#> 
#> 
#> [[89]]
#> [[89]]$Sepal.Length
#> [1] 5.6
#> 
#> [[89]]$Sepal.Width
#> [1] 3
#> 
#> 
#> [[90]]
#> [[90]]$Sepal.Length
#> [1] 5.5
#> 
#> [[90]]$Sepal.Width
#> [1] 2.5
#> 
#> 
#> [[91]]
#> [[91]]$Sepal.Length
#> [1] 5.5
#> 
#> [[91]]$Sepal.Width
#> [1] 2.6
#> 
#> 
#> [[92]]
#> [[92]]$Sepal.Length
#> [1] 6.1
#> 
#> [[92]]$Sepal.Width
#> [1] 3
#> 
#> 
#> [[93]]
#> [[93]]$Sepal.Length
#> [1] 5.8
#> 
#> [[93]]$Sepal.Width
#> [1] 2.6
#> 
#> 
#> [[94]]
#> [[94]]$Sepal.Length
#> [1] 5
#> 
#> [[94]]$Sepal.Width
#> [1] 2.3
#> 
#> 
#> [[95]]
#> [[95]]$Sepal.Length
#> [1] 5.6
#> 
#> [[95]]$Sepal.Width
#> [1] 2.7
#> 
#> 
#> [[96]]
#> [[96]]$Sepal.Length
#> [1] 5.7
#> 
#> [[96]]$Sepal.Width
#> [1] 3
#> 
#> 
#> [[97]]
#> [[97]]$Sepal.Length
#> [1] 5.7
#> 
#> [[97]]$Sepal.Width
#> [1] 2.9
#> 
#> 
#> [[98]]
#> [[98]]$Sepal.Length
#> [1] 6.2
#> 
#> [[98]]$Sepal.Width
#> [1] 2.9
#> 
#> 
#> [[99]]
#> [[99]]$Sepal.Length
#> [1] 5.1
#> 
#> [[99]]$Sepal.Width
#> [1] 2.5
#> 
#> 
#> [[100]]
#> [[100]]$Sepal.Length
#> [1] 5.7
#> 
#> [[100]]$Sepal.Width
#> [1] 2.8
#> 
#> 
#> [[101]]
#> [[101]]$Sepal.Length
#> [1] 6.3
#> 
#> [[101]]$Sepal.Width
#> [1] 3.3
#> 
#> 
#> [[102]]
#> [[102]]$Sepal.Length
#> [1] 5.8
#> 
#> [[102]]$Sepal.Width
#> [1] 2.7
#> 
#> 
#> [[103]]
#> [[103]]$Sepal.Length
#> [1] 7.1
#> 
#> [[103]]$Sepal.Width
#> [1] 3
#> 
#> 
#> [[104]]
#> [[104]]$Sepal.Length
#> [1] 6.3
#> 
#> [[104]]$Sepal.Width
#> [1] 2.9
#> 
#> 
#> [[105]]
#> [[105]]$Sepal.Length
#> [1] 6.5
#> 
#> [[105]]$Sepal.Width
#> [1] 3
#> 
#> 
#> [[106]]
#> [[106]]$Sepal.Length
#> [1] 7.6
#> 
#> [[106]]$Sepal.Width
#> [1] 3
#> 
#> 
#> [[107]]
#> [[107]]$Sepal.Length
#> [1] 4.9
#> 
#> [[107]]$Sepal.Width
#> [1] 2.5
#> 
#> 
#> [[108]]
#> [[108]]$Sepal.Length
#> [1] 7.3
#> 
#> [[108]]$Sepal.Width
#> [1] 2.9
#> 
#> 
#> [[109]]
#> [[109]]$Sepal.Length
#> [1] 6.7
#> 
#> [[109]]$Sepal.Width
#> [1] 2.5
#> 
#> 
#> [[110]]
#> [[110]]$Sepal.Length
#> [1] 7.2
#> 
#> [[110]]$Sepal.Width
#> [1] 3.6
#> 
#> 
#> [[111]]
#> [[111]]$Sepal.Length
#> [1] 6.5
#> 
#> [[111]]$Sepal.Width
#> [1] 3.2
#> 
#> 
#> [[112]]
#> [[112]]$Sepal.Length
#> [1] 6.4
#> 
#> [[112]]$Sepal.Width
#> [1] 2.7
#> 
#> 
#> [[113]]
#> [[113]]$Sepal.Length
#> [1] 6.8
#> 
#> [[113]]$Sepal.Width
#> [1] 3
#> 
#> 
#> [[114]]
#> [[114]]$Sepal.Length
#> [1] 5.7
#> 
#> [[114]]$Sepal.Width
#> [1] 2.5
#> 
#> 
#> [[115]]
#> [[115]]$Sepal.Length
#> [1] 5.8
#> 
#> [[115]]$Sepal.Width
#> [1] 2.8
#> 
#> 
#> [[116]]
#> [[116]]$Sepal.Length
#> [1] 6.4
#> 
#> [[116]]$Sepal.Width
#> [1] 3.2
#> 
#> 
#> [[117]]
#> [[117]]$Sepal.Length
#> [1] 6.5
#> 
#> [[117]]$Sepal.Width
#> [1] 3
#> 
#> 
#> [[118]]
#> [[118]]$Sepal.Length
#> [1] 7.7
#> 
#> [[118]]$Sepal.Width
#> [1] 3.8
#> 
#> 
#> [[119]]
#> [[119]]$Sepal.Length
#> [1] 7.7
#> 
#> [[119]]$Sepal.Width
#> [1] 2.6
#> 
#> 
#> [[120]]
#> [[120]]$Sepal.Length
#> [1] 6
#> 
#> [[120]]$Sepal.Width
#> [1] 2.2
#> 
#> 
#> [[121]]
#> [[121]]$Sepal.Length
#> [1] 6.9
#> 
#> [[121]]$Sepal.Width
#> [1] 3.2
#> 
#> 
#> [[122]]
#> [[122]]$Sepal.Length
#> [1] 5.6
#> 
#> [[122]]$Sepal.Width
#> [1] 2.8
#> 
#> 
#> [[123]]
#> [[123]]$Sepal.Length
#> [1] 7.7
#> 
#> [[123]]$Sepal.Width
#> [1] 2.8
#> 
#> 
#> [[124]]
#> [[124]]$Sepal.Length
#> [1] 6.3
#> 
#> [[124]]$Sepal.Width
#> [1] 2.7
#> 
#> 
#> [[125]]
#> [[125]]$Sepal.Length
#> [1] 6.7
#> 
#> [[125]]$Sepal.Width
#> [1] 3.3
#> 
#> 
#> [[126]]
#> [[126]]$Sepal.Length
#> [1] 7.2
#> 
#> [[126]]$Sepal.Width
#> [1] 3.2
#> 
#> 
#> [[127]]
#> [[127]]$Sepal.Length
#> [1] 6.2
#> 
#> [[127]]$Sepal.Width
#> [1] 2.8
#> 
#> 
#> [[128]]
#> [[128]]$Sepal.Length
#> [1] 6.1
#> 
#> [[128]]$Sepal.Width
#> [1] 3
#> 
#> 
#> [[129]]
#> [[129]]$Sepal.Length
#> [1] 6.4
#> 
#> [[129]]$Sepal.Width
#> [1] 2.8
#> 
#> 
#> [[130]]
#> [[130]]$Sepal.Length
#> [1] 7.2
#> 
#> [[130]]$Sepal.Width
#> [1] 3
#> 
#> 
#> [[131]]
#> [[131]]$Sepal.Length
#> [1] 7.4
#> 
#> [[131]]$Sepal.Width
#> [1] 2.8
#> 
#> 
#> [[132]]
#> [[132]]$Sepal.Length
#> [1] 7.9
#> 
#> [[132]]$Sepal.Width
#> [1] 3.8
#> 
#> 
#> [[133]]
#> [[133]]$Sepal.Length
#> [1] 6.4
#> 
#> [[133]]$Sepal.Width
#> [1] 2.8
#> 
#> 
#> [[134]]
#> [[134]]$Sepal.Length
#> [1] 6.3
#> 
#> [[134]]$Sepal.Width
#> [1] 2.8
#> 
#> 
#> [[135]]
#> [[135]]$Sepal.Length
#> [1] 6.1
#> 
#> [[135]]$Sepal.Width
#> [1] 2.6
#> 
#> 
#> [[136]]
#> [[136]]$Sepal.Length
#> [1] 7.7
#> 
#> [[136]]$Sepal.Width
#> [1] 3
#> 
#> 
#> [[137]]
#> [[137]]$Sepal.Length
#> [1] 6.3
#> 
#> [[137]]$Sepal.Width
#> [1] 3.4
#> 
#> 
#> [[138]]
#> [[138]]$Sepal.Length
#> [1] 6.4
#> 
#> [[138]]$Sepal.Width
#> [1] 3.1
#> 
#> 
#> [[139]]
#> [[139]]$Sepal.Length
#> [1] 6
#> 
#> [[139]]$Sepal.Width
#> [1] 3
#> 
#> 
#> [[140]]
#> [[140]]$Sepal.Length
#> [1] 6.9
#> 
#> [[140]]$Sepal.Width
#> [1] 3.1
#> 
#> 
#> [[141]]
#> [[141]]$Sepal.Length
#> [1] 6.7
#> 
#> [[141]]$Sepal.Width
#> [1] 3.1
#> 
#> 
#> [[142]]
#> [[142]]$Sepal.Length
#> [1] 6.9
#> 
#> [[142]]$Sepal.Width
#> [1] 3.1
#> 
#> 
#> [[143]]
#> [[143]]$Sepal.Length
#> [1] 5.8
#> 
#> [[143]]$Sepal.Width
#> [1] 2.7
#> 
#> 
#> [[144]]
#> [[144]]$Sepal.Length
#> [1] 6.8
#> 
#> [[144]]$Sepal.Width
#> [1] 3.2
#> 
#> 
#> [[145]]
#> [[145]]$Sepal.Length
#> [1] 6.7
#> 
#> [[145]]$Sepal.Width
#> [1] 3.3
#> 
#> 
#> [[146]]
#> [[146]]$Sepal.Length
#> [1] 6.7
#> 
#> [[146]]$Sepal.Width
#> [1] 3
#> 
#> 
#> [[147]]
#> [[147]]$Sepal.Length
#> [1] 6.3
#> 
#> [[147]]$Sepal.Width
#> [1] 2.5
#> 
#> 
#> [[148]]
#> [[148]]$Sepal.Length
#> [1] 6.5
#> 
#> [[148]]$Sepal.Width
#> [1] 3
#> 
#> 
#> [[149]]
#> [[149]]$Sepal.Length
#> [1] 6.2
#> 
#> [[149]]$Sepal.Width
#> [1] 3.4
#> 
#> 
#> [[150]]
#> [[150]]$Sepal.Length
#> [1] 5.9
#> 
#> [[150]]$Sepal.Width
#> [1] 3
#> 
#> 

# Use custom names

parse_df(iris[, 1:2], add_names = c("x", "y"))

#> [[1]]
#> [[1]]$x
#> [1] 5.1
#> 
#> [[1]]$y
#> [1] 3.5
#> 
#> 
#> [[2]]
#> [[2]]$x
#> [1] 4.9
#> 
#> [[2]]$y
#> [1] 3
#> 
#> 
#> [[3]]
#> [[3]]$x
#> [1] 4.7
#> 
#> [[3]]$y
#> [1] 3.2
#> 
#> 
#> [[4]]
#> [[4]]$x
#> [1] 4.6
#> 
#> [[4]]$y
#> [1] 3.1
#> 
#> 
#> [[5]]
#> [[5]]$x
#> [1] 5
#> 
#> [[5]]$y
#> [1] 3.6
#> 
#> 
#> [[6]]
#> [[6]]$x
#> [1] 5.4
#> 
#> [[6]]$y
#> [1] 3.9
#> 
#> 
#> [[7]]
#> [[7]]$x
#> [1] 4.6
#> 
#> [[7]]$y
#> [1] 3.4
#> 
#> 
#> [[8]]
#> [[8]]$x
#> [1] 5
#> 
#> [[8]]$y
#> [1] 3.4
#> 
#> 
#> [[9]]
#> [[9]]$x
#> [1] 4.4
#> 
#> [[9]]$y
#> [1] 2.9
#> 
#> 
#> [[10]]
#> [[10]]$x
#> [1] 4.9
#> 
#> [[10]]$y
#> [1] 3.1
#> 
#> 
#> [[11]]
#> [[11]]$x
#> [1] 5.4
#> 
#> [[11]]$y
#> [1] 3.7
#> 
#> 
#> [[12]]
#> [[12]]$x
#> [1] 4.8
#> 
#> [[12]]$y
#> [1] 3.4
#> 
#> 
#> [[13]]
#> [[13]]$x
#> [1] 4.8
#> 
#> [[13]]$y
#> [1] 3
#> 
#> 
#> [[14]]
#> [[14]]$x
#> [1] 4.3
#> 
#> [[14]]$y
#> [1] 3
#> 
#> 
#> [[15]]
#> [[15]]$x
#> [1] 5.8
#> 
#> [[15]]$y
#> [1] 4
#> 
#> 
#> [[16]]
#> [[16]]$x
#> [1] 5.7
#> 
#> [[16]]$y
#> [1] 4.4
#> 
#> 
#> [[17]]
#> [[17]]$x
#> [1] 5.4
#> 
#> [[17]]$y
#> [1] 3.9
#> 
#> 
#> [[18]]
#> [[18]]$x
#> [1] 5.1
#> 
#> [[18]]$y
#> [1] 3.5
#> 
#> 
#> [[19]]
#> [[19]]$x
#> [1] 5.7
#> 
#> [[19]]$y
#> [1] 3.8
#> 
#> 
#> [[20]]
#> [[20]]$x
#> [1] 5.1
#> 
#> [[20]]$y
#> [1] 3.8
#> 
#> 
#> [[21]]
#> [[21]]$x
#> [1] 5.4
#> 
#> [[21]]$y
#> [1] 3.4
#> 
#> 
#> [[22]]
#> [[22]]$x
#> [1] 5.1
#> 
#> [[22]]$y
#> [1] 3.7
#> 
#> 
#> [[23]]
#> [[23]]$x
#> [1] 4.6
#> 
#> [[23]]$y
#> [1] 3.6
#> 
#> 
#> [[24]]
#> [[24]]$x
#> [1] 5.1
#> 
#> [[24]]$y
#> [1] 3.3
#> 
#> 
#> [[25]]
#> [[25]]$x
#> [1] 4.8
#> 
#> [[25]]$y
#> [1] 3.4
#> 
#> 
#> [[26]]
#> [[26]]$x
#> [1] 5
#> 
#> [[26]]$y
#> [1] 3
#> 
#> 
#> [[27]]
#> [[27]]$x
#> [1] 5
#> 
#> [[27]]$y
#> [1] 3.4
#> 
#> 
#> [[28]]
#> [[28]]$x
#> [1] 5.2
#> 
#> [[28]]$y
#> [1] 3.5
#> 
#> 
#> [[29]]
#> [[29]]$x
#> [1] 5.2
#> 
#> [[29]]$y
#> [1] 3.4
#> 
#> 
#> [[30]]
#> [[30]]$x
#> [1] 4.7
#> 
#> [[30]]$y
#> [1] 3.2
#> 
#> 
#> [[31]]
#> [[31]]$x
#> [1] 4.8
#> 
#> [[31]]$y
#> [1] 3.1
#> 
#> 
#> [[32]]
#> [[32]]$x
#> [1] 5.4
#> 
#> [[32]]$y
#> [1] 3.4
#> 
#> 
#> [[33]]
#> [[33]]$x
#> [1] 5.2
#> 
#> [[33]]$y
#> [1] 4.1
#> 
#> 
#> [[34]]
#> [[34]]$x
#> [1] 5.5
#> 
#> [[34]]$y
#> [1] 4.2
#> 
#> 
#> [[35]]
#> [[35]]$x
#> [1] 4.9
#> 
#> [[35]]$y
#> [1] 3.1
#> 
#> 
#> [[36]]
#> [[36]]$x
#> [1] 5
#> 
#> [[36]]$y
#> [1] 3.2
#> 
#> 
#> [[37]]
#> [[37]]$x
#> [1] 5.5
#> 
#> [[37]]$y
#> [1] 3.5
#> 
#> 
#> [[38]]
#> [[38]]$x
#> [1] 4.9
#> 
#> [[38]]$y
#> [1] 3.6
#> 
#> 
#> [[39]]
#> [[39]]$x
#> [1] 4.4
#> 
#> [[39]]$y
#> [1] 3
#> 
#> 
#> [[40]]
#> [[40]]$x
#> [1] 5.1
#> 
#> [[40]]$y
#> [1] 3.4
#> 
#> 
#> [[41]]
#> [[41]]$x
#> [1] 5
#> 
#> [[41]]$y
#> [1] 3.5
#> 
#> 
#> [[42]]
#> [[42]]$x
#> [1] 4.5
#> 
#> [[42]]$y
#> [1] 2.3
#> 
#> 
#> [[43]]
#> [[43]]$x
#> [1] 4.4
#> 
#> [[43]]$y
#> [1] 3.2
#> 
#> 
#> [[44]]
#> [[44]]$x
#> [1] 5
#> 
#> [[44]]$y
#> [1] 3.5
#> 
#> 
#> [[45]]
#> [[45]]$x
#> [1] 5.1
#> 
#> [[45]]$y
#> [1] 3.8
#> 
#> 
#> [[46]]
#> [[46]]$x
#> [1] 4.8
#> 
#> [[46]]$y
#> [1] 3
#> 
#> 
#> [[47]]
#> [[47]]$x
#> [1] 5.1
#> 
#> [[47]]$y
#> [1] 3.8
#> 
#> 
#> [[48]]
#> [[48]]$x
#> [1] 4.6
#> 
#> [[48]]$y
#> [1] 3.2
#> 
#> 
#> [[49]]
#> [[49]]$x
#> [1] 5.3
#> 
#> [[49]]$y
#> [1] 3.7
#> 
#> 
#> [[50]]
#> [[50]]$x
#> [1] 5
#> 
#> [[50]]$y
#> [1] 3.3
#> 
#> 
#> [[51]]
#> [[51]]$x
#> [1] 7
#> 
#> [[51]]$y
#> [1] 3.2
#> 
#> 
#> [[52]]
#> [[52]]$x
#> [1] 6.4
#> 
#> [[52]]$y
#> [1] 3.2
#> 
#> 
#> [[53]]
#> [[53]]$x
#> [1] 6.9
#> 
#> [[53]]$y
#> [1] 3.1
#> 
#> 
#> [[54]]
#> [[54]]$x
#> [1] 5.5
#> 
#> [[54]]$y
#> [1] 2.3
#> 
#> 
#> [[55]]
#> [[55]]$x
#> [1] 6.5
#> 
#> [[55]]$y
#> [1] 2.8
#> 
#> 
#> [[56]]
#> [[56]]$x
#> [1] 5.7
#> 
#> [[56]]$y
#> [1] 2.8
#> 
#> 
#> [[57]]
#> [[57]]$x
#> [1] 6.3
#> 
#> [[57]]$y
#> [1] 3.3
#> 
#> 
#> [[58]]
#> [[58]]$x
#> [1] 4.9
#> 
#> [[58]]$y
#> [1] 2.4
#> 
#> 
#> [[59]]
#> [[59]]$x
#> [1] 6.6
#> 
#> [[59]]$y
#> [1] 2.9
#> 
#> 
#> [[60]]
#> [[60]]$x
#> [1] 5.2
#> 
#> [[60]]$y
#> [1] 2.7
#> 
#> 
#> [[61]]
#> [[61]]$x
#> [1] 5
#> 
#> [[61]]$y
#> [1] 2
#> 
#> 
#> [[62]]
#> [[62]]$x
#> [1] 5.9
#> 
#> [[62]]$y
#> [1] 3
#> 
#> 
#> [[63]]
#> [[63]]$x
#> [1] 6
#> 
#> [[63]]$y
#> [1] 2.2
#> 
#> 
#> [[64]]
#> [[64]]$x
#> [1] 6.1
#> 
#> [[64]]$y
#> [1] 2.9
#> 
#> 
#> [[65]]
#> [[65]]$x
#> [1] 5.6
#> 
#> [[65]]$y
#> [1] 2.9
#> 
#> 
#> [[66]]
#> [[66]]$x
#> [1] 6.7
#> 
#> [[66]]$y
#> [1] 3.1
#> 
#> 
#> [[67]]
#> [[67]]$x
#> [1] 5.6
#> 
#> [[67]]$y
#> [1] 3
#> 
#> 
#> [[68]]
#> [[68]]$x
#> [1] 5.8
#> 
#> [[68]]$y
#> [1] 2.7
#> 
#> 
#> [[69]]
#> [[69]]$x
#> [1] 6.2
#> 
#> [[69]]$y
#> [1] 2.2
#> 
#> 
#> [[70]]
#> [[70]]$x
#> [1] 5.6
#> 
#> [[70]]$y
#> [1] 2.5
#> 
#> 
#> [[71]]
#> [[71]]$x
#> [1] 5.9
#> 
#> [[71]]$y
#> [1] 3.2
#> 
#> 
#> [[72]]
#> [[72]]$x
#> [1] 6.1
#> 
#> [[72]]$y
#> [1] 2.8
#> 
#> 
#> [[73]]
#> [[73]]$x
#> [1] 6.3
#> 
#> [[73]]$y
#> [1] 2.5
#> 
#> 
#> [[74]]
#> [[74]]$x
#> [1] 6.1
#> 
#> [[74]]$y
#> [1] 2.8
#> 
#> 
#> [[75]]
#> [[75]]$x
#> [1] 6.4
#> 
#> [[75]]$y
#> [1] 2.9
#> 
#> 
#> [[76]]
#> [[76]]$x
#> [1] 6.6
#> 
#> [[76]]$y
#> [1] 3
#> 
#> 
#> [[77]]
#> [[77]]$x
#> [1] 6.8
#> 
#> [[77]]$y
#> [1] 2.8
#> 
#> 
#> [[78]]
#> [[78]]$x
#> [1] 6.7
#> 
#> [[78]]$y
#> [1] 3
#> 
#> 
#> [[79]]
#> [[79]]$x
#> [1] 6
#> 
#> [[79]]$y
#> [1] 2.9
#> 
#> 
#> [[80]]
#> [[80]]$x
#> [1] 5.7
#> 
#> [[80]]$y
#> [1] 2.6
#> 
#> 
#> [[81]]
#> [[81]]$x
#> [1] 5.5
#> 
#> [[81]]$y
#> [1] 2.4
#> 
#> 
#> [[82]]
#> [[82]]$x
#> [1] 5.5
#> 
#> [[82]]$y
#> [1] 2.4
#> 
#> 
#> [[83]]
#> [[83]]$x
#> [1] 5.8
#> 
#> [[83]]$y
#> [1] 2.7
#> 
#> 
#> [[84]]
#> [[84]]$x
#> [1] 6
#> 
#> [[84]]$y
#> [1] 2.7
#> 
#> 
#> [[85]]
#> [[85]]$x
#> [1] 5.4
#> 
#> [[85]]$y
#> [1] 3
#> 
#> 
#> [[86]]
#> [[86]]$x
#> [1] 6
#> 
#> [[86]]$y
#> [1] 3.4
#> 
#> 
#> [[87]]
#> [[87]]$x
#> [1] 6.7
#> 
#> [[87]]$y
#> [1] 3.1
#> 
#> 
#> [[88]]
#> [[88]]$x
#> [1] 6.3
#> 
#> [[88]]$y
#> [1] 2.3
#> 
#> 
#> [[89]]
#> [[89]]$x
#> [1] 5.6
#> 
#> [[89]]$y
#> [1] 3
#> 
#> 
#> [[90]]
#> [[90]]$x
#> [1] 5.5
#> 
#> [[90]]$y
#> [1] 2.5
#> 
#> 
#> [[91]]
#> [[91]]$x
#> [1] 5.5
#> 
#> [[91]]$y
#> [1] 2.6
#> 
#> 
#> [[92]]
#> [[92]]$x
#> [1] 6.1
#> 
#> [[92]]$y
#> [1] 3
#> 
#> 
#> [[93]]
#> [[93]]$x
#> [1] 5.8
#> 
#> [[93]]$y
#> [1] 2.6
#> 
#> 
#> [[94]]
#> [[94]]$x
#> [1] 5
#> 
#> [[94]]$y
#> [1] 2.3
#> 
#> 
#> [[95]]
#> [[95]]$x
#> [1] 5.6
#> 
#> [[95]]$y
#> [1] 2.7
#> 
#> 
#> [[96]]
#> [[96]]$x
#> [1] 5.7
#> 
#> [[96]]$y
#> [1] 3
#> 
#> 
#> [[97]]
#> [[97]]$x
#> [1] 5.7
#> 
#> [[97]]$y
#> [1] 2.9
#> 
#> 
#> [[98]]
#> [[98]]$x
#> [1] 6.2
#> 
#> [[98]]$y
#> [1] 2.9
#> 
#> 
#> [[99]]
#> [[99]]$x
#> [1] 5.1
#> 
#> [[99]]$y
#> [1] 2.5
#> 
#> 
#> [[100]]
#> [[100]]$x
#> [1] 5.7
#> 
#> [[100]]$y
#> [1] 2.8
#> 
#> 
#> [[101]]
#> [[101]]$x
#> [1] 6.3
#> 
#> [[101]]$y
#> [1] 3.3
#> 
#> 
#> [[102]]
#> [[102]]$x
#> [1] 5.8
#> 
#> [[102]]$y
#> [1] 2.7
#> 
#> 
#> [[103]]
#> [[103]]$x
#> [1] 7.1
#> 
#> [[103]]$y
#> [1] 3
#> 
#> 
#> [[104]]
#> [[104]]$x
#> [1] 6.3
#> 
#> [[104]]$y
#> [1] 2.9
#> 
#> 
#> [[105]]
#> [[105]]$x
#> [1] 6.5
#> 
#> [[105]]$y
#> [1] 3
#> 
#> 
#> [[106]]
#> [[106]]$x
#> [1] 7.6
#> 
#> [[106]]$y
#> [1] 3
#> 
#> 
#> [[107]]
#> [[107]]$x
#> [1] 4.9
#> 
#> [[107]]$y
#> [1] 2.5
#> 
#> 
#> [[108]]
#> [[108]]$x
#> [1] 7.3
#> 
#> [[108]]$y
#> [1] 2.9
#> 
#> 
#> [[109]]
#> [[109]]$x
#> [1] 6.7
#> 
#> [[109]]$y
#> [1] 2.5
#> 
#> 
#> [[110]]
#> [[110]]$x
#> [1] 7.2
#> 
#> [[110]]$y
#> [1] 3.6
#> 
#> 
#> [[111]]
#> [[111]]$x
#> [1] 6.5
#> 
#> [[111]]$y
#> [1] 3.2
#> 
#> 
#> [[112]]
#> [[112]]$x
#> [1] 6.4
#> 
#> [[112]]$y
#> [1] 2.7
#> 
#> 
#> [[113]]
#> [[113]]$x
#> [1] 6.8
#> 
#> [[113]]$y
#> [1] 3
#> 
#> 
#> [[114]]
#> [[114]]$x
#> [1] 5.7
#> 
#> [[114]]$y
#> [1] 2.5
#> 
#> 
#> [[115]]
#> [[115]]$x
#> [1] 5.8
#> 
#> [[115]]$y
#> [1] 2.8
#> 
#> 
#> [[116]]
#> [[116]]$x
#> [1] 6.4
#> 
#> [[116]]$y
#> [1] 3.2
#> 
#> 
#> [[117]]
#> [[117]]$x
#> [1] 6.5
#> 
#> [[117]]$y
#> [1] 3
#> 
#> 
#> [[118]]
#> [[118]]$x
#> [1] 7.7
#> 
#> [[118]]$y
#> [1] 3.8
#> 
#> 
#> [[119]]
#> [[119]]$x
#> [1] 7.7
#> 
#> [[119]]$y
#> [1] 2.6
#> 
#> 
#> [[120]]
#> [[120]]$x
#> [1] 6
#> 
#> [[120]]$y
#> [1] 2.2
#> 
#> 
#> [[121]]
#> [[121]]$x
#> [1] 6.9
#> 
#> [[121]]$y
#> [1] 3.2
#> 
#> 
#> [[122]]
#> [[122]]$x
#> [1] 5.6
#> 
#> [[122]]$y
#> [1] 2.8
#> 
#> 
#> [[123]]
#> [[123]]$x
#> [1] 7.7
#> 
#> [[123]]$y
#> [1] 2.8
#> 
#> 
#> [[124]]
#> [[124]]$x
#> [1] 6.3
#> 
#> [[124]]$y
#> [1] 2.7
#> 
#> 
#> [[125]]
#> [[125]]$x
#> [1] 6.7
#> 
#> [[125]]$y
#> [1] 3.3
#> 
#> 
#> [[126]]
#> [[126]]$x
#> [1] 7.2
#> 
#> [[126]]$y
#> [1] 3.2
#> 
#> 
#> [[127]]
#> [[127]]$x
#> [1] 6.2
#> 
#> [[127]]$y
#> [1] 2.8
#> 
#> 
#> [[128]]
#> [[128]]$x
#> [1] 6.1
#> 
#> [[128]]$y
#> [1] 3
#> 
#> 
#> [[129]]
#> [[129]]$x
#> [1] 6.4
#> 
#> [[129]]$y
#> [1] 2.8
#> 
#> 
#> [[130]]
#> [[130]]$x
#> [1] 7.2
#> 
#> [[130]]$y
#> [1] 3
#> 
#> 
#> [[131]]
#> [[131]]$x
#> [1] 7.4
#> 
#> [[131]]$y
#> [1] 2.8
#> 
#> 
#> [[132]]
#> [[132]]$x
#> [1] 7.9
#> 
#> [[132]]$y
#> [1] 3.8
#> 
#> 
#> [[133]]
#> [[133]]$x
#> [1] 6.4
#> 
#> [[133]]$y
#> [1] 2.8
#> 
#> 
#> [[134]]
#> [[134]]$x
#> [1] 6.3
#> 
#> [[134]]$y
#> [1] 2.8
#> 
#> 
#> [[135]]
#> [[135]]$x
#> [1] 6.1
#> 
#> [[135]]$y
#> [1] 2.6
#> 
#> 
#> [[136]]
#> [[136]]$x
#> [1] 7.7
#> 
#> [[136]]$y
#> [1] 3
#> 
#> 
#> [[137]]
#> [[137]]$x
#> [1] 6.3
#> 
#> [[137]]$y
#> [1] 3.4
#> 
#> 
#> [[138]]
#> [[138]]$x
#> [1] 6.4
#> 
#> [[138]]$y
#> [1] 3.1
#> 
#> 
#> [[139]]
#> [[139]]$x
#> [1] 6
#> 
#> [[139]]$y
#> [1] 3
#> 
#> 
#> [[140]]
#> [[140]]$x
#> [1] 6.9
#> 
#> [[140]]$y
#> [1] 3.1
#> 
#> 
#> [[141]]
#> [[141]]$x
#> [1] 6.7
#> 
#> [[141]]$y
#> [1] 3.1
#> 
#> 
#> [[142]]
#> [[142]]$x
#> [1] 6.9
#> 
#> [[142]]$y
#> [1] 3.1
#> 
#> 
#> [[143]]
#> [[143]]$x
#> [1] 5.8
#> 
#> [[143]]$y
#> [1] 2.7
#> 
#> 
#> [[144]]
#> [[144]]$x
#> [1] 6.8
#> 
#> [[144]]$y
#> [1] 3.2
#> 
#> 
#> [[145]]
#> [[145]]$x
#> [1] 6.7
#> 
#> [[145]]$y
#> [1] 3.3
#> 
#> 
#> [[146]]
#> [[146]]$x
#> [1] 6.7
#> 
#> [[146]]$y
#> [1] 3
#> 
#> 
#> [[147]]
#> [[147]]$x
#> [1] 6.3
#> 
#> [[147]]$y
#> [1] 2.5
#> 
#> 
#> [[148]]
#> [[148]]$x
#> [1] 6.5
#> 
#> [[148]]$y
#> [1] 3
#> 
#> 
#> [[149]]
#> [[149]]$x
#> [1] 6.2
#> 
#> [[149]]$y
#> [1] 3.4
#> 
#> 
#> [[150]]
#> [[150]]$x
#> [1] 5.9
#> 
#> [[150]]$y
#> [1] 3
#> 
#>