Radial bar options

Use these options in ax_plotOptions().

Usage

radialBar_opts(
  size = NULL,
  inverseOrder = NULL,
  startAngle = NULL,
  endAngle = NULL,
  offsetX = NULL,
  offsetY = NULL,
  hollow = NULL,
  track = NULL,
  dataLabels = NULL,
  ...
)

Arguments

size

Numeric. Manual size of the radialBars instead of calculating automatically from default height / width.

inverseOrder

Logical. Whether to make the first value of series innermost or outermost.

startAngle

Numeric. Angle from which the radialBars should start.

endAngle

Numeric. Angle to which the radialBars should end. The sum of the startAngle and endAngle should not exceed 360.

offsetX

Numeric. Sets the left offset for radialBars.

offsetY

Numeric. Sets the top offset for radialBars.

hollow

List.

track

List.

dataLabels

List.

...

Additional parameters.

Value

A list of options that can be used in ax_plotOptions().

Note

See https://apexcharts.com/docs/options/plotoptions/radialbar/.

Examples

apexchart() %>% 
  ax_chart(type = "radialBar") %>% 
  ax_plotOptions(
    radialBar = radialBar_opts(
      startAngle = -135,
      endAngle = 135,
      dataLabels = list(
        name = list(
          fontSize = "16px",
          # color = undefined,
          offsetY = 120
        ),
        value = list(
          offsetY = 76,
          fontSize = "22px",
          # color = undefined,
          formatter = htmlwidgets::JS("function (val) {return val + '%';}")
        )
      )
    )
  ) %>% 
  ax_stroke(dashArray = 4) %>% 
  ax_series(70) %>% 
  ax_labels("Indicator")